requirements.txt
gunicorn
Django==4.1.1
djangorestframework==3.14.0
 
fpdf==1.7.2
reportlab==3.6.11
whitenoise==6.2.0
psycopg2==2.9.3
PyPDF2==2.11.0
django-cors-headers==3.13.0


